requirements.txt
alembic==1.11.1
anyio==3.7.0
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
Django==4.2.3
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.99.1
Flask==2.3.2
fonttools==4.40.0
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
libcst==1.0.1
Mako==1.2.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
passlib==1.7.4
Pillow==9.5.0
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.10
Pygments==2.15.1
pymongo==4.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rich==13.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
starlette==0.27.0
strawberry-graphql==0.194.4
typer==0.9.0
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uuid==1.30
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.6
